requirements.txt
pypdf
chromadb
pytest
PyYAML~=6.0.2
python-dotenv~=1.0.1
langchain~=0.3.0
langchain-chroma
langchain-core~=0.3.5
langchain-openai~=0.2.0
langchain-text-splitters~=0.3.0
langchain-community~=0.3.0